fix issues caused or discovered by code style tweaks
