+++ mod/metric_answer_lookup/lib/answer/answer_class_methods.rb
@@ -108,14 +108,14 @@ class Answer
 
     def latest_answer_card metric_id, company_id
       a_id = where(metric_id: metric_id, company_id: company_id,
-                   latest: true).pluck(:answer_id)
+                   latest: true).pluck(:answer_id).first
       a_id && Card.fetch(a_id)
     end
 
     def latest_year metric_id, company_id
       where(metric_id: metric_id,
             company_id: company_id,
-            latest: true).pluck(:year)
+            latest: true).pluck(:year).first
     end
 
     def answered? metric_id, company_id
