+++ mod/metrics/set/abstract/01_metric_child.rb
@@ -38,7 +38,7 @@ def metric_was
 end
 
 def metric_card
-  Card.fetch(metric) || (generation == 1 && left)
+  Card.fetch(metric) || (generation == 1 && left) || nil
   # FIXME: hack to make it work on new cards
 end
